Archive tab postponed.
CreateOrderDialog overhaul. Added more info.
TaskMonitor tableView auto-updates every 5 seconds.

todo: Archive tab

+++ src/main/java/com/sunmyoung/tasktracker/controllers/TaskMonitorController.java
@@ -2,61 +2,65 @@ package com.sunmyoung.tasktracker.controllers;
 
 import com.sunmyoung.tasktracker.Launcher;
 import com.sunmyoung.tasktracker.controllers.dialogControllers.CreateOrderDialogController;
+import com.sunmyoung.tasktracker.controllers.dialogControllers.CreateOrderDialogControllerV2;
 import com.sunmyoung.tasktracker.pojos.Task;
 import com.sunmyoung.tasktracker.repositories.Database;
 import com.sunmyoung.tasktracker.repositories.TaskRepository;
+import javafx.application.Platform;
 import javafx.beans.property.ReadOnlyObjectWrapper;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.control.cell.TextFieldTableCell;
-import javafx.scene.input.MouseButton;
-import javafx.scene.input.MouseEvent;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
-import java.util.List;
-import java.util.Optional;
+import java.time.LocalDate;
+import java.util.*;
 
 public class TaskMonitorController {
     @Getter
-    private ObservableList<Task> data;
+    private ObservableList<Task> tasksObservableList = FXCollections.observableArrayList();
 
     @FXML
-    private TableColumn<Task, String> deadlineCol,
-            orderDateCol,
+    private TableColumn<Task, LocalDate>
+            deadlineCol,
+            dateInCol;
+
+    @FXML
+    private TableColumn<Task, Integer> countCol;
+
+    @FXML
+    private TableColumn<Task, String>
             orderCol,
             clientCol,
-            countCol,
             sizeCol,
             meshCol,
             combiCol,
             typeCol,
             tensioningCol,
             coatingCol,
-            washingCol,
+            packagingCol,
             exposureCol;
 
     @FXML
     @Getter
     private TableView<Task> tableView;
 
-
-    @FXML
-    public void refresh(ActionEvent event) {
-        loadData();
-    }
-
     @FXML
     void createNewTask(ActionEvent event) {
         createOrderDialog();
+    }
+
+    @FXML
+    void refresh(ActionEvent event) {
+        refreshTable();
     }
 
     @FXML
@@ -78,33 +82,46 @@ public class TaskMonitorController {
         initTableView();
         loadData();
         activateStatusCells();
+        startAutoRefresh(5000);
     }
 
     public void loadData() {
         List<Task> taskList = TaskRepository.getUnfinishedTasks();
+        if (taskList == null) {
+            taskList = new ArrayList<>();
+        }
         System.out.println(taskList);
 
-        data = FXCollections.observableArrayList(taskList);
-        tableView.setItems(data);
+        tasksObservableList.clear();
+        tasksObservableList.addAll(taskList);
+    }
+
+    /**
+     * Manually refresh the tasks table.
+     */
+    private void refreshTable() {
+        tasksObservableList.clear();
+        loadData();
     }
 
     private void initTableView() {
+        tableView.setItems(tasksObservableList);
+
         //Single cell selection
         tableView.getSelectionModel().cellSelectionEnabledProperty().set(true);
         tableView.setEditable(true);
 
         orderCol.setCellValueFactory(order -> new ReadOnlyObjectWrapper<>(Integer.toString(tableView.getItems().indexOf(order.getValue()) + 1)));
         deadlineCol.setCellValueFactory(new PropertyValueFactory<>("deadline"));
-        orderDateCol.setCellValueFactory(new PropertyValueFactory<>("orderDateStr"));
+        dateInCol.setCellValueFactory(new PropertyValueFactory<>("dateIn"));
         clientCol.setCellValueFactory(new PropertyValueFactory<>("client"));
         countCol.setCellValueFactory(new PropertyValueFactory<>("count"));
-        sizeCol.setCellValueFactory(new PropertyValueFactory<>("size"));
+        sizeCol.setCellValueFactory(new PropertyValueFactory<>("frameSize"));
         meshCol.setCellValueFactory(new PropertyValueFactory<>("mesh"));
         combiCol.setCellValueFactory(new PropertyValueFactory<>("combi"));
-        typeCol.setCellValueFactory(new PropertyValueFactory<>("aluminum"));
         tensioningCol.setCellValueFactory(new PropertyValueFactory<>("tensioning"));
         coatingCol.setCellValueFactory(new PropertyValueFactory<>("coating"));
-        washingCol.setCellValueFactory(new PropertyValueFactory<>("washing"));
+        packagingCol.setCellValueFactory(new PropertyValueFactory<>("packaging"));
         exposureCol.setCellValueFactory(new PropertyValueFactory<>("exposure"));
     }
 
@@ -171,9 +188,9 @@ public class TaskMonitorController {
     }
 
     private void activateWashingCol() {
-        washingCol.setEditable(true);
-        washingCol.setCellFactory(TextFieldTableCell.forTableColumn());
-        washingCol.setOnEditCommit(event -> {
+        packagingCol.setEditable(true);
+        packagingCol.setCellFactory(TextFieldTableCell.forTableColumn());
+        packagingCol.setOnEditCommit(event -> {
             Task selectedTask = event.getTableView().getItems().get(event.getTablePosition().getRow());
             String newValue = event.getNewValue();
             if (!newValue.equals(event.getOldValue())){
@@ -183,7 +200,7 @@ public class TaskMonitorController {
                     transaction = session.beginTransaction();
                     selectedTask = session.createQuery("from Task t where t.id = :id", Task.class).setParameter("id", selectedTask.getId()).uniqueResult();
                     //here
-                    selectedTask.setWashing(newValue);
+                    selectedTask.setPackaging(newValue);
                     transaction.commit();
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -229,12 +246,12 @@ public class TaskMonitorController {
         Task selectedTask = tableView.getSelectionModel().getSelectedItem();
         long taskId = selectedTask.getId();
 
-        FXMLLoader fxmlLoader = new FXMLLoader(Launcher.class.getResource("dialogs/createOrderDialog.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(Launcher.class.getResource("dialogs/createOrderDialogV2.fxml"));
         DialogPane dialogPane = fxmlLoader.load();
         Dialog<ButtonType> dialog = new Dialog<>();
         dialog.setDialogPane(dialogPane);
 
-        CreateOrderDialogController controller = fxmlLoader.getController();
+        CreateOrderDialogControllerV2 controller = fxmlLoader.getController();
 
         Session session = Database.getSessionFactory().openSession();
         Transaction transaction = null;
@@ -247,12 +264,18 @@ public class TaskMonitorController {
 
             Optional<ButtonType> clickedButton = dialog.showAndWait();
             if(clickedButton.isPresent() && clickedButton.get() == ButtonType.OK) {
-                controller.readFields(task);
-                session.createQuery("delete from Subtask where task = null").executeUpdate();
-            }
+                //subtasksChanged is triggered if addSubtask or removeSubtask buttons were pressed.
+                if (controller.isSubtasksChanged()) {
+                    task.getSubtasks().forEach(subtask -> subtask.setTask(null));
+                    controller.readFields(task);
+                    session.createQuery("delete from Subtask where task = null").executeUpdate();
+                } else {
+                    controller.readFields(task);
+                }
 
-            transaction.commit();
-            loadData();
+                transaction.commit();
+                loadData();
+            }
         } catch (Exception e) {
             e.printStackTrace();
             if(transaction != null)
@@ -264,18 +287,41 @@ public class TaskMonitorController {
 
     @SneakyThrows
     private void createOrderDialog() {
-        FXMLLoader fxmlLoader = new FXMLLoader(Launcher.class.getResource("dialogs/createOrderDialog.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(Launcher.class.getResource("dialogs/createOrderDialogV2.fxml"));
         DialogPane dialogPane = fxmlLoader.load();
         Dialog<ButtonType> dialog = new Dialog<>();
         dialog.setDialogPane(dialogPane);
 
-        CreateOrderDialogController controller = fxmlLoader.getController();
+        CreateOrderDialogControllerV2 controller = fxmlLoader.getController();
 
         Optional<ButtonType> clickedButton = dialog.showAndWait();
         if (clickedButton.isPresent() && clickedButton.get() == ButtonType.OK) {
-            controller.createTask();
+            Task task = new Task();
+            controller.readFields(task);
+            TaskRepository.save(task);
             loadData();
         }
     }
 
+    /**
+     * @param delay - pause time between table updates in milliseconds.
+     */
+    private void startAutoRefresh(int delay) {
+        Thread thread = new Thread(new Runnable() {
+            @Override
+            @SneakyThrows
+            public void run() {
+                while (true) {
+                    Thread.sleep(delay);
+                    Platform.runLater(() -> {
+                        List<Task> taskList = TaskRepository.getUnfinishedTasks();
+                        tasksObservableList.clear();
+                        tasksObservableList.addAll(taskList);
+                        System.out.println("Refreshed");
+                    });
+                }
+            }
+        });
+        thread.start();
+    }
 }
